add DOT
