+++ index/src/assets/mod.rs
@@ -1,2 +1,4 @@
 mod acala;
+mod phala;
 pub use acala::*;
+pub use phala::*;
